requirements.txt
# 웹 크롤링 관련
selenium>=4.15.0
requests>=2.31.0
python-dotenv>=1.0.0

# 데이터 분석 및 Parquet 변환
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0